fix: refactor and fix failing tests
