+++ src/app.module.ts
@@ -25,7 +25,7 @@ import { CachesModule, DateTimeModule } from './common';
 import { EthereumListenerModule } from './listeners/ethereum-listener/ethereum-listener.module';
 import { TransactionModule } from './endpoints/transaction/transaction.module';
 import { SpecializationModule } from './endpoints/category/specialization/specialization.module';
-import { NotificationEndpointModule } from './endpoints/notification-endpoint/notification-endpoint.module';
+import { NotificationEndpointModule } from './common/modules/notification-endpoint/notification-endpoint.module';
 
 require('dotenv').config(); // eslint-disable-line
