+++ src/endpoints/email/email.module.ts
@@ -4,11 +4,16 @@ import {
   MailModule,
   SubstrateModule,
 } from '../../common';
-import { NotificationModule } from '../notification/notification.module';
+import { NotificationModule } from '../../common/modules/notification/notification.module';
 import { EmailEndpointController } from './email.controller';
 
 @Module({
-  imports: [MailModule, SubstrateModule, EmailNotificationModule, NotificationModule],
+  imports: [
+    MailModule,
+    SubstrateModule,
+    EmailNotificationModule,
+    NotificationModule,
+  ],
   controllers: [EmailEndpointController],
 })
 export class EmailEndpointModule {}
