+++ test/e2e/endpoints/email.spec.ts
@@ -9,6 +9,7 @@ import {
   ProcessEnvModule,
   SubstrateModule,
   SubstrateService,
+  NotificationModule,
 } from '../../../src/common';
 import request from 'supertest';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -44,6 +45,7 @@ describe('Email Controller (e2e)', () => {
         SubstrateModule,
         EmailNotificationModule,
         ProcessEnvModule,
+        NotificationModule,
         EmailEndpointModule,
       ],
     }).compile();
@@ -63,6 +65,7 @@ describe('Email Controller (e2e)', () => {
       'sendLabRegistrationEmail',
     );
     sendLabRegistrationEmailSpy.mockImplementation(() => Promise.resolve(true));
+
     const result = await request(server)
       .post(`/email/registered-lab/${substrateService.pair.address}`)
       .send();
