+++ test/unit/endpoints/email/email.controller.spec.ts
@@ -16,13 +16,12 @@ import { Response } from 'express';
 import { Test, TestingModule } from '@nestjs/testing';
 import * as labQuery from '@debionetwork/polkadot-provider/lib/query/labs';
 import { when } from 'jest-when';
-import { NotificationService } from '../../../../src/endpoints/notification/notification.service';
+import { NotificationService } from '../../../../src/common/modules/notification/notification.service';
 
 describe('Email Controller', () => {
   let emailEndpointControllerMock: EmailEndpointController;
   let mailerManageMock: MockType<MailerManager>;
   let substrateServiceMock: MockType<SubstrateService>;
-  let notificationService: NotificationService// eslint-disable-line
   const EMAILS = 'email';
 
   class ProcessEnvProxyMock {
@@ -62,7 +61,6 @@ describe('Email Controller', () => {
     );
     mailerManageMock = module.get(MailerManager);
     substrateServiceMock = module.get(SubstrateService);
-    notificationService = module.get(NotificationService);// eslint-disable-line
   });
 
   it('should be defined', () => {
